# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API documentation for anoweb.",
        "title": "anoweb API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/static/upload-image": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "static"
                ],
                "summary": "Upload image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/statistics/daily-active": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Daily active users",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Days (1-365)",
                        "name": "days",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.DailyActiveUsersPoint"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/statistics/overall": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Overall statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OverallStatisticsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/statistics/streak": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Current user streak",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.StreakResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/statistics/users-over-time": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Visitors over time",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 24,
                        "description": "Hours (1-168)",
                        "name": "hours",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.FansOverTimePoint"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.OverallStatisticsResponse": {
            "type": "object",
            "properties": {
                "active_users_today": {
                    "type": "integer"
                },
                "guest_visitors_24h": {
                    "type": "integer"
                },
                "guest_visitors_ever": {
                    "type": "integer"
                },
                "registered_visitors_24h": {
                    "type": "integer"
                },
                "registered_visitors_ever": {
                    "type": "integer"
                },
                "total_hours": {
                    "type": "number"
                },
                "total_users": {
                    "type": "integer"
                },
                "unique_visitors_24h": {
                    "type": "integer"
                },
                "unique_visitors_ever": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.StreakResponse": {
            "type": "object",
            "properties": {
                "streak": {
                    "type": "integer"
                }
            }
        },
...
        "repositories.DailyActiveUsersPoint": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "repositories.FansOverTimePoint": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "hour": {
                    "type": "string"
                }
            }
        }
    }
}